Quick example on how to use mox
