+++ farmbot_ext/mix.exs
@@ -30,6 +30,7 @@ defmodule FarmbotExt.MixProject do
       {:hackney, "~> 1.14"},
       {:uuid, "~> 1.1"},
       {:amqp, "1.1.1"},
+      {:mox, "~> 0.5", only: :test},
       {:excoveralls, "~> 0.10", only: [:test], targets: [:host]},
       {:dialyxir, "~> 1.0.0-rc.3", only: [:dev], targets: [:host], runtime: false},
       {:ex_doc, "~> 0.19", only: [:dev], targets: [:host], runtime: false}
