+++ farmbot_ext/test/test_helper.exs
@@ -1,1 +1,2 @@
+Mox.defmock(MockPreloader, for: FarmbotExt.API.Preloader)
 ExUnit.start()
